api/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.1.0
pydantic>=2.4.2
openai>=1.78.1 